Added abstract class for wiki-wide features swo we can restrict wiki-wide stubbing
